L' calculation now uses Bruno EdgeCorrection().
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@1483 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/funcs.h
@@ -244,7 +244,7 @@ extern int   SetAutocuts(struct threshold_s *thresh, struct plan7_s *hmm);
  */
 extern void  P7Config(struct plan7_s *hmm, enum p7_algmode mode);
 extern void  P7ReconfigLength(struct plan7_s *hmm, int L);
-extern float P7FinalBitscore(struct plan7_s *hmm, int sc, int L);
+extern float P7FinalBitscore(struct plan7_s *hmm, float sc, int L);
 
 
 /* from modelmakers.c
